- use floating point when computing # of cylinders.  The expression   hdsize*1024*1024/16/63/512 was prone to overflowing in 32-bit integer math.   Thanks to Henry Mohn III <animatorhmiii@netscape.net> for sending a patch.

+++ misc/bximage.c
@@ -244,7 +244,7 @@ int main()
     int hdsize, cyl, heads=16, spt=63;
     if (ask_int ("\nEnter the hard disk size in megabytes, between 1 and 32255\n[10] ", 1, 32255, 10, &hdsize) < 0)
       fatal (EOF_ERR);
-    cyl = hdsize*1024*1024/16/63/512;
+    cyl = (int) (hdsize*1024.0*1024.0/16.0/63.0/512.0);
     assert (cyl < 65536);
     sectors = cyl*heads*spt;
     printf ("\nI will create a hard disk image with\n");
